--- a/src/ReusableTasks/ActionWorkItem.cs
+++ b/src/ReusableTasks/ActionWorkItem.cs
@@ -1,19 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using System.Threading;
 
 namespace ReusableTasks
 {
-#if NET5_0_OR_GREATER || NETCOREAPP3_0_OR_GREATER
+#if !NETSTANDARD2_0 && !NETSTANDARD2_1
     class ActionWorkItem : IThreadPoolWorkItem
     {
         static readonly Action EmptyAction = () => { };
         static readonly Stack<ActionWorkItem> Cache = new Stack<ActionWorkItem> ();
         static readonly SimpleSpinLock CacheLock = new SimpleSpinLock ();
 
-        public Action Continuation { get; private set; } = EmptyAction;
+        public object Continuation { get; private set; } = EmptyAction;
 
-        public static ActionWorkItem GetOrCreate (Action action)
+        public static ActionWorkItem GetOrCreate (object action)
         {
             using (CacheLock.Enter ()) {
                 if (Cache.Count == 0) {
@@ -28,12 +30,11 @@ namespace ReusableTasks
 
         public void Execute ()
         {
-            var continuation = Continuation;
+            ResultHolder.Invoker (Continuation);
             Continuation = EmptyAction;
 
             using (CacheLock.Enter ())
                 Cache.Push (this);
-            continuation ();
         }
     }
 #endif