… src/ReusableTasks/System.Runtime.CompilerServices/StateMachineWithActionCache.cs
… version.txt